requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
python-dotenv
scikit-learn
joblib
pandas
